Changes to sectors.db structure and the subsequent reprocussions
Changed `sectors.db` table to also contain ['cap','market_cap'] by pushing `capitalization()` function from `update_arkfund()` to `update_sectors()`.

This was done because it is better for the longevity of the project ... ['sector','cap','market_cap'] are values that remain relatively static and can be stored in a db for quick reference instead of a finviz.API call for millions of row entries.

This change resolved an issue with `change_in_portfolio()` where ticker=['NNDM'] was designated as "Mid" cap on 1/13 but suddenly dropped to "Small" cap on 1/15 and the system reported the position as "closed" because it could no longer find a Mid-cap company called NNDM.

Subsequent changes were made update_arkfund() to account for this new method of obtaining ['cap','market_cap'] for the `ARKFund.db` database.

Additionally `update_capitalization()` was pushed to the main function `update_arkfund()` and programmed to trigger exclusively on Friday.

The new `update_capitalizations()` now updates BOTH `ARKFund.db` and `sectors.db` and processes the data significantly faster (took 40s for a few days of data before, now only takes ~5s with minimal Finviz.API calls because it can simply grab the data from unqiue tickers of `sectors.db` instead of an API call for each row of `ARKFund.db`)

Lastly, `update_arkfund()` has been declared & made the "main" function. Previously it purposed kind of like a main function, but still required some other functions to run after it. Now it is a one-in-all wrapper for setting up & updating the database. Printouts have thus been updated to provide better clarity on which sub-routine is currently executing & what step along the process is occuring by updating printouts for ['update_sectors()','backup_data()','update_capitalizations()']