Created ARK.py + changes to update_capitalizations() computation
With the project moving to an applet very soon (via Dash/Plotly) the functions need to be accessible through the (to be created) app.py stand-alone script.

`ARK.py` will contain all the functions written in the jupyter notebook condensed into a .py script -- that was it can be loaded in via "from ARK import *" ... I say "will contain" because it currently contains the necessary functions to run the database but is missing some functionality like "ticker_lookup()" as that has no use until I can build the Applet to house the plots.

Changes to `update_capitalization()` the method by which "ARKFund.db" was updated. Originally it was computed by iterating through all n-rows of `ARKFund.db`. Now it utilizes a dictionary derived from `sectors.db` to map and generate ['cap','market_cap'] in the `ARKFund.db` df near instantaneously.

Minor changes here and there:
- Added "manual_trigger" argument to "update_capitalization()" so it can trigger manually.
- Updated `update_arkfund()` to also contain this "manual_trigger" argument so it can be passed down to the subroutine.